Agency controller
*

Fixing agency controller
Using service instead repository on controller
Fixin service
Changing id from int to long
Creating repository method to update the entity but not the password

*

+++ src/main/java/br/com/vsgdev/emotionsRoute/repository/EquipamentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.vsgdev.emotionsRoute.model.Equipament;
 
 @Repository
-public interface EquipamentRepository extends CrudRepository<Equipament, Integer>{
+public interface EquipamentRepository extends CrudRepository<Equipament, Long>{
 
 }
