+++ src/main/java/br/com/vsgdev/emotionsRoute/service/EquipamentService.java
@@ -5,12 +5,12 @@ import br.com.vsgdev.emotionsRoute.model.Equipament;
 
 public interface EquipamentService {
 
-	Equipament get(int id) throws NotFoundEntity;
+	Equipament get( Long id ) throws NotFoundEntity;
 
-	void delete(int id) throws NotFoundEntity;
+	void delete( Long id ) throws NotFoundEntity;
 
-	Equipament put(Equipament equipament) throws NotFoundEntity;
+	Equipament put( Equipament equipament ) throws NotFoundEntity;
 
-	Equipament save(Equipament equipament);
+	Equipament save( Equipament equipament );
 
 }
