+++ src/main/java/br/com/vsgdev/emotionsRoute/serviceImpl/EquipamentServiceImpl.java
@@ -17,31 +17,31 @@ public class EquipamentServiceImpl implements EquipamentService {
 	private EquipamentRepository eqRep;
 
 	@Override
-	public Equipament get(int id) throws NotFoundEntity {
-		Optional<Equipament> result = eqRep.findById(id);
-		if (result.isPresent()) {
+	public Equipament get( Long id ) throws NotFoundEntity {
+		Optional<Equipament> result = eqRep.findById( id );
+		if ( result.isPresent() ) {
 			return result.get();
 		} else {
-			throw new NotFoundEntity("id", String.valueOf(id));
+			throw new NotFoundEntity( "id", String.valueOf( id ) );
 		}
 	}
 
 	@Override
-	public void delete(int id) throws NotFoundEntity {
-		eqRep.delete(get(id));
+	public void delete( Long id ) throws NotFoundEntity {
+		eqRep.delete( get( id ) );
 
 	}
 
 	@Override
-	public Equipament put(Equipament equipament) throws NotFoundEntity {
-		get(equipament.getId());
-		return eqRep.save(equipament);
+	public Equipament put( Equipament equipament ) throws NotFoundEntity {
+		get( equipament.getId() );
+		return eqRep.save( equipament );
 	}
 
 	@Override
-	public Equipament save(Equipament equipament) {
-		if (equipament.getId() == 0) {
-			return eqRep.save(equipament);
+	public Equipament save( Equipament equipament ) {
+		if ( equipament.getId() == 0 ) {
+			return eqRep.save( equipament );
 		} else {
 			// may throw an exception
 			return null;
